refactor: makeAutoTicket 함수 분리
